+++ src/managed/OpenLiveWriter.HtmlEditor/HtmlEditorControl.cs
@@ -4570,24 +4570,13 @@ namespace OpenLiveWriter.HtmlEditor
                     IUndoUnit undoUnit = CreateUndoUnit();
                     using (undoUnit)
                     {
-                        const string preOpen = "<pre>";
-                        const string preClose = "</pre>";
-                        htmlText = RemoveCode(htmlText, preOpen, preClose);
-
-                        string styledHtml = $"{preOpen}{htmlText}{preClose}";
+                        var styledHtml = CodeHighlighter.StyledHtml(htmlText);
                         InsertHtml(range.Start, range.End, styledHtml);
                         // commit the change
                         undoUnit.Commit();
                     }
                 }
             }
-        }
-
-        private static string RemoveCode(string htmlText, string preOpen, string preClose)
-        {
-            htmlText = htmlText.Replace(preOpen, String.Empty).Replace(preClose, String.Empty);
-            htmlText = htmlText.Replace(preOpen.ToUpper(), String.Empty).Replace(preClose.ToUpper(), String.Empty);
-            return htmlText;
         }
 
         bool IHtmlEditorCommandSource.CanInsertLink
